# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.CreateIncomingWebhookTokenRequest:
    properties:
      default_priority:
        type: string
      default_tags:
        items:
          type: string
        type: array
      integration_type:
        $ref: '#/definitions/domain.IncomingWebhookIntegrationType'
      name:
        type: string
    required:
    - integration_type
    - name
    type: object
...
  domain.CreateWebhookEndpointRequest:
    properties:
      alert_acknowledged:
        type: boolean
      alert_closed:
        type: boolean
      alert_created:
        type: boolean
      alert_escalated:
        type: boolean
      alert_updated:
        type: boolean
      enabled:
        type: boolean
      headers:
        additionalProperties:
          type: string
        type: object
      incident_created:
        type: boolean
      incident_resolved:
        type: boolean
      incident_updated:
        type: boolean
      max_retries:
        type: integer
      name:
        type: string
      retry_delay_seconds:
        type: integer
      timeout_seconds:
        type: integer
      url:
        type: string
    required:
    - name
    - url
    type: object
...
  domain.IncomingWebhookIntegrationType:
    enum:
    - generic
    - prometheus
    - grafana
    - datadog
    type: string
    x-enum-varnames:
    - IncomingWebhookGeneric
    - IncomingWebhookPrometheus
    - IncomingWebhookGrafana
    - IncomingWebhookDatadog
  domain.IncomingWebhookToken:
    properties:
      created_at:
        type: string
      default_priority:
        type: string
      default_tags:
        items:
          type: string
        type: array
      enabled:
        type: boolean
      id:
        type: string
      integration_type:
        $ref: '#/definitions/domain.IncomingWebhookIntegrationType'
      last_used_at:
        type: string
      name:
        type: string
      organization_id:
        type: string
      request_count:
        type: integer
      token:
        type: string
      updated_at:
        type: string
    type: object
...
  domain.UpdateWebhookEndpointRequest:
    properties:
      alert_acknowledged:
        type: boolean
      alert_closed:
        type: boolean
      alert_created:
        type: boolean
      alert_escalated:
        type: boolean
      alert_updated:
        type: boolean
      enabled:
        type: boolean
      headers:
        additionalProperties:
          type: string
        type: object
      incident_created:
        type: boolean
      incident_resolved:
        type: boolean
      incident_updated:
        type: boolean
      max_retries:
        type: integer
      name:
        type: string
      retry_delay_seconds:
        type: integer
      timeout_seconds:
        type: integer
      url:
        type: string
    type: object
...
  domain.WebhookEndpoint:
    properties:
      alert_acknowledged:
        type: boolean
      alert_closed:
        type: boolean
      alert_created:
        description: Event filters
        type: boolean
      alert_escalated:
        type: boolean
      alert_updated:
        type: boolean
      created_at:
        type: string
      enabled:
        type: boolean
      headers:
        additionalProperties:
          type: string
        description: HTTP configuration
        type: object
      id:
        type: string
      incident_created:
        type: boolean
      incident_resolved:
        type: boolean
      incident_updated:
        type: boolean
      max_retries:
        description: Retry configuration
        type: integer
      name:
        type: string
      organization_id:
        type: string
      retry_delay_seconds:
        type: integer
      timeout_seconds:
        type: integer
      updated_at:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pulsar.io
    name: API Support
    url: https://github.com/nmn3m/pulsar
  description: Pulsar is an incident management and alerting platform. This API provides
    endpoints for managing alerts, incidents, teams, schedules, escalation policies,
    notifications, and webhooks.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pulsar API
  version: "1.0"
paths:
...
  /webhooks/endpoints:
    get:
      consumes:
      - application/json
      description: List all outgoing webhook endpoints for the organization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.WebhookEndpoint'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List webhook endpoints
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Create a new outgoing webhook endpoint
      parameters:
      - description: Create endpoint request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateWebhookEndpointRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.WebhookEndpoint'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a webhook endpoint
      tags:
      - Webhooks
  /webhooks/endpoints/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a webhook endpoint by ID
      parameters:
      - description: Endpoint ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a webhook endpoint
      tags:
      - Webhooks
    get:
      consumes:
      - application/json
      description: Get a webhook endpoint by ID
      parameters:
      - description: Endpoint ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.WebhookEndpoint'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a webhook endpoint
      tags:
      - Webhooks
    patch:
      consumes:
      - application/json
      description: Update a webhook endpoint by ID
      parameters:
      - description: Endpoint ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Update endpoint request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateWebhookEndpointRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.WebhookEndpoint'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a webhook endpoint
      tags:
      - Webhooks
  /webhooks/incoming:
    get:
      consumes:
      - application/json
      description: List all incoming webhook tokens for the organization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.IncomingWebhookToken'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List incoming webhook tokens
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Create a new incoming webhook token for receiving webhooks from
        external sources
      parameters:
      - description: Create token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateIncomingWebhookTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.IncomingWebhookToken'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create an incoming webhook token
      tags:
      - Webhooks
  /webhooks/incoming/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an incoming webhook token by ID
      parameters:
      - description: Token ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an incoming webhook token
      tags:
      - Webhooks
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"